@@ -10,7 +10,7 @@ import android.util.Log;
 
 public class AlarmReceiver extends BroadcastReceiver {
 
-    private final String TAG = "AlarmReceiver";
+    private static final String TAG = "AlarmReceiver";
 
     @Override
     public void onReceive(Context context, Intent intent) {
