@@ -1,22 +1,27 @@
 package com.example.yink.amadeus;
 
 import android.content.SharedPreferences;
-import android.os.Build;
 import android.os.Bundle;
 import android.preference.PreferenceFragment;
-import android.support.annotation.RequiresApi;
+import android.widget.Toast;
 
 /**
  * Created by Yink on 05.03.2017.
  */
 
-@RequiresApi(api = Build.VERSION_CODES.HONEYCOMB)
-public class SettingsFragment extends PreferenceFragment {
-    SharedPreferences sharedPreferences;
+public class SettingsFragment extends PreferenceFragment
+        implements SharedPreferences.OnSharedPreferenceChangeListener {
 
     @Override
     public void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         addPreferencesFromResource(R.xml.preferences);
+    }
+
+    @Override
+    public void onSharedPreferenceChanged(SharedPreferences sharedPreferences, String key) {
+        if (key.equals("lang")) {
+            Toast.makeText(getActivity(), getString(R.string.pref_app_restart_required), Toast.LENGTH_SHORT).show();
+        }
     }
 }
